Deprecated capability "expiredRead"
This capability does not make sense as it was previusly for informing
that items if a storage adapter can be read if the TTL will be increased
or set to infinity but this only works on storage adapters with capability
"static_ttl" and increasing the TTL on the same time means that the item
is no longer expired. So it's basically providing the same information as
"static_ttl" but from a wrong PoV.

The capability is now returning the information provided by "static_ttl"
and will trigger a E_USER_DEPRECATED message.